requirement.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==4.0
django-allauth==0.61.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-spectacular==0.27.2
idna==3.7
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1